elxr v0.1.0 Elxr.Enum_operations

Make an enum_operations module, it should have the following methods: count, reverse, filter, concat. You need to build the same logic as the Enum module does for this methods (just for list). The point of this exercise is to create this basic functions yourself.

Functions

concat(enumerable)

  iex(1)> Elxr.Enum_operations.concat([[123, 32], ["hola", :ok]]) 
  [123, 32, "hola", :ok]
  [123, 32, "hola", :ok]

count(enumerable)

  iex(1)> Elxr.Enum_operations.count([12, 123, 345, 456, 76])                         
  El enumerable pasado tiene 5 elementos
  :ok

filter(enumerable, fun)

  iex(1)> Elxr.Enum_operations.filter([1, 2, 3, 4, 5], fn x -> x <= 4 end)
  [1, 2, 3, 4]
  [1, 2, 3, 4]

reverse(enumerable)

  iex(1)> Elxr.Enum_operations.reverse [12, 123, 345, "hola", %{"nombre" => "pablo"}] 
  [%{"nombre" => "pablo"}, "hola", 345, 123, 12]
  [%{"nombre" => "pablo"}, "hola", 345, 123, 12]