elxr v0.1.0 Elxr.Fizz_buzz

Make a fizz_buzz module and write a function called fuzzle that takes three arguments. If the first two are zero, return “FizzBuzz”. If the first is zero, return “Fizz”. If the second is zero return “Buzz”. Otherwise return the third argument. You don't need to use recursion here, create 2 methods, one using cond do, and the other one case do.

Functions

fuzzle(a, b, c)

iex(4)> Elxr.Fizz_buzz.fuzzle(0, 0, 3)    
"FizzBuzz"

fuzzle2(lista)

iex(3)> Elxr.Fizz_buzz.fuzzle2([1, 0, 3]) 
"Buzz"